update preview and on frontpage
