+++ apps/theme/app/themebuilder/_components/ThemePages.module.css
@@ -1,0 +1,13 @@
+.panel {
+  background-color: var(--ds-color-neutral-background-default);
+  border: 1px solid var(--ds-color-neutral-border-subtle);
+  border-radius: 8px;
+  margin-top: 28px;
+  width: 100%;
+  max-width: 1390px;
+  padding: var(--ds-size-6);
+}
+
+.colorsContainer {
+  overflow: auto;
+}
