+++ apps/theme/app/themebuilder/_components/ThemePages.tsx
@@ -1,0 +1,109 @@
+'use client';
+
+import cl from 'clsx/lite';
+import { useEffect, useRef } from 'react';
+import { ColorContrasts, ColorPreview, ColorTokens } from '../../../components';
+import { Colors } from '../../../components/Colors/Colors';
+import { OverviewComponents } from '../../../components/OverviewComponents/OverviewComponents';
+import { useThemeStore } from '../../../store';
+import classes from './ThemePages.module.css';
+
+export const ThemePages = () => {
+  const baseBorderRadius = useThemeStore((state) => state.baseBorderRadius);
+  const themeTab = useThemeStore((state) => state.themeTab);
+  const appearance = useThemeStore((state) => state.appearance);
+
+  const containerRef = useRef<HTMLDivElement>(null);
+
+  useEffect(() => {
+    // we need to set these properties on the preview element because they are immutable on :root
+    for (const key in borderRadiuses) {
+      const borderRadius = borderRadiuses[key as keyof typeof borderRadiuses];
+
+      if (containerRef.current) {
+        containerRef.current.style.setProperty(
+          borderRadius.variable,
+          borderRadius.value,
+        );
+      }
+    }
+  }, []);
+
+  useEffect(() => {
+    if (containerRef.current) {
+      containerRef.current.style.setProperty(
+        '--ds-border-radius-base',
+        `${baseBorderRadius / 16}rem`,
+      );
+    }
+  }, [baseBorderRadius]);
+
+  return (
+    <>
+      {themeTab === 'overview' && (
+        <div className={classes.panel} data-color-scheme={appearance}>
+          <OverviewComponents />
+        </div>
+      )}
+
+      {themeTab === 'colorsystem' && (
+        <>
+          <div
+            className={cl(classes.panel, classes.colorsContainer)}
+            data-color-scheme={appearance}
+          >
+            <Colors />
+          </div>
+
+          <div className={classes.panel} data-color-scheme={appearance}>
+            <ColorPreview />
+          </div>
+          <div className={classes.panel} data-color-scheme={appearance}>
+            <ColorTokens />
+          </div>
+
+          <div className={classes.panel} data-color-scheme={appearance}>
+            <ColorContrasts />
+          </div>
+        </>
+      )}
+    </>
+  );
+};
+
+// TODO get this token data from @digdir/designsystemet (use json from --preview or something)
+const borderRadiuses = {
+  sm: {
+    name: 'sm',
+    value:
+      'min(var(--ds-border-radius-base)*0.5,var(--ds-border-radius-scale))',
+    variable: '--ds-border-radius-sm',
+  },
+  md: {
+    name: 'md',
+    value: 'min(var(--ds-border-radius-base),var(--ds-border-radius-scale)*2)',
+    variable: '--ds-border-radius-md',
+  },
+  lg: {
+    name: 'lg',
+    value:
+      'min(var(--ds-border-radius-base)*2,var(--ds-border-radius-scale)*5)',
+    variable: '--ds-border-radius-lg',
+  },
+  xl: {
+    name: 'xl',
+    value:
+      'min(var(--ds-border-radius-base)*3,var(--ds-border-radius-scale)*7)',
+    variable: '--ds-border-radius-xl',
+  },
+  default: {
+    name: 'default',
+    value: 'var(--ds-border-radius-base)',
+    variable: '--ds-border-radius-default',
+  },
+  full: {
+    name: 'full',
+    value: '624.9375rem',
+    variable: '--ds-border-radius-full',
+  },
+};
