+++ apps/theme/app/themebuilder/page.module.css
@@ -9,10 +9,6 @@
   flex-grow: 500;
 }
 
-.colorsContainer {
-  overflow: auto;
-}
-
 .sideBarContainer {
   width: 420px;
 }
@@ -22,16 +18,6 @@
   justify-content: space-between;
   flex-wrap: wrap;
   gap: 24px;
-}
-
-.panel {
-  background-color: var(--ds-color-neutral-background-default);
-  border: 1px solid var(--ds-color-neutral-border-subtle);
-  border-radius: 8px;
-  margin-top: 28px;
-  width: 100%;
-  max-width: 1390px;
-  padding: var(--ds-size-6);
 }
 
 .hide {
