+++ apps/theme/app/themebuilder/page.tsx
@@ -1,22 +1,10 @@
 'use client';
-import cl from 'clsx/lite';
-import {
-  BorderRadius,
-  ColorContrasts,
-  ColorPreview,
-  ColorTokens,
-  Colors,
-  Header,
-  Sidebar,
-} from '../../components';
-import { useThemeStore } from '../../store';
+import { Header, Sidebar } from '../../components';
+import { ThemePages } from './_components/ThemePages';
 import { useThemeParams } from './_utils/useThemeParams';
 import classes from './page.module.css';
 
-export default function Home() {
-  const themeTab = useThemeStore((state) => state.themeTab);
-  const appearance = useThemeStore((state) => state.appearance);
-
+export default function Page() {
   /* For theme params */
   useThemeParams();
 
@@ -25,33 +13,7 @@ export default function Home() {
       <Header />
       <div className={classes.container}>
         <div className={classes.content}>
-          {themeTab === 'overview' && (
-            <div className={classes.panel} data-color-scheme={appearance}>
-              <BorderRadius />
-            </div>
-          )}
-
-          {themeTab === 'colorsystem' && (
-            <>
-              <div
-                className={cl(classes.panel, classes.colorsContainer)}
-                data-color-scheme={appearance}
-              >
-                <Colors />
-              </div>
-
-              <div className={classes.panel} data-color-scheme={appearance}>
-                <ColorPreview />
-              </div>
-              <div className={classes.panel} data-color-scheme={appearance}>
-                <ColorTokens />
-              </div>
-
-              <div className={classes.panel} data-color-scheme={appearance}>
-                <ColorContrasts />
-              </div>
-            </>
-          )}
+          <ThemePages />
         </div>
         <div className={classes.sideBarContainer}>
           <Sidebar />
