+++ apps/theme/components/OverviewComponents/OverviewComponents.module.css
@@ -28,9 +28,9 @@
 }
 
 .inner {
-  display: grid;
-  grid-template-columns: repeat(2, 1fr);
-  gap: 28px;
+  display: flex;
+  gap: var(--ds-size-6);
+  flex-wrap: wrap;
 }
 
 .card {
@@ -60,11 +60,6 @@
   margin-top: 4px;
 }
 
-.users {
-  display: flex;
-  flex-direction: column;
-}
-
 .img {
   width: 100%;
   border-radius: var(--ds-border-radius-md);
@@ -87,6 +82,11 @@
   font-size: 16px;
 }
 
+.users {
+  display: flex;
+  flex-direction: column;
+}
+
 .user {
   display: flex;
   align-items: center;
@@ -102,12 +102,6 @@
   margin-bottom: 0;
 }
 
-.userImg {
-  height: 44px;
-  width: 44px;
-  border-radius: var(--ds-border-radius-default);
-}
-
 .userRole {
   font-size: 15px;
   font-weight: 500;
@@ -115,14 +109,9 @@
   margin-bottom: -1px;
 }
 
-.userName {
-  font-size: 16px;
-}
-
-.userBtn {
-  margin-left: auto;
-}
-
-.test {
-  border-radius: var(--ds-border-radius-xl);
+.darkCard {
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  flex-grow: 1;
 }
