+++ apps/theme/components/OverviewComponents/SettingsCard/SettingsCard.tsx
@@ -5,9 +5,15 @@ import {
   Radio,
   Switch,
   ValidationMessage,
+  useRadioGroup,
 } from '@digdir/designsystemet-react';
 import classes from './SettingsCard.module.css';
 export const SettingsCard = () => {
+  const { getRadioProps } = useRadioGroup({
+    name: 'my-group',
+    value: 'sjokolade',
+  });
+
   return (
     <div>
       <Heading data-size='2xs'>Innstillinger</Heading>
@@ -33,22 +39,17 @@ export const SettingsCard = () => {
 
       <Fieldset data-size='sm' className={classes.radioGroup}>
         <Fieldset.Legend>Visnigsmodus</Fieldset.Legend>
+        <Radio label='Vanilje' {...getRadioProps('vanilje')} />
         <Radio
-          ref={function Ms() {}}
-          label='Vanilje'
-          name='my-group'
-          onChange={function Ms() {}}
-          value='vanilje'
-          checked
-        />
-        <Radio
-          ref={function Ms() {}}
           label='Jordbær'
-          name='my-group'
-          onChange={function Ms() {}}
-          value='jordbær'
+          description='Jordbær er best'
+          {...getRadioProps('jordbær')}
         />
-
+        <Radio label='Sjokolade' {...getRadioProps('sjokolade')} />
+        <Radio
+          label='Jeg spiser ikke iskrem'
+          {...getRadioProps('spiser-ikke-is')}
+        />
         <ValidationMessage hidden id=':re:' />
       </Fieldset>
     </div>
