+++ apps/theme/components/Previews/Previews.module.css
@@ -1,59 +1,12 @@
 .toolbar {
   display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin: 36px 0 24px;
-}
-
-.menuItem[aria-disabled] {
-  cursor: not-allowed;
-  opacity: 0.6;
-  background-color: transparent !important;
-}
-
-.toggles {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  gap: 8px;
-}
-
-.toggle {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  border: none;
-  background-color: transparent;
-  padding: 9px 16px;
-  font-size: 15px;
-  border-radius: 500px;
-  color: var(--ds-color-accent-text-default);
-}
-
-.toggle:hover {
-  cursor: pointer;
-  background-color: var(--ds-color-accent-surface-hover);
-}
-
-.active {
-  background-color: var(--ds-color-accent-surface-default);
-}
-
-.toggle img {
-  height: 16px;
+  gap: var(--ds-size-4);
+  margin: var(--ds-size-4) 0;
+  margin-top: var(--ds-size-22);
 }
 
 .preview {
-  --background: var(--neutral-2);
-  --foreground: var(--neutral-1);
-  width: 100%;
-  border-radius: 12px;
-  display: flex;
-  gap: 32px;
-}
-
-.preview[data-color-scheme='dark'],
-.preview[data-color-scheme='contrast'] {
-  --background: var(--neutral-1);
-  --foreground: var(--neutral-2);
+  background-color: var(--ds-color-neutral-surface-default);
+  padding: var(--ds-size-4);
+  border-radius: var(--ds-border-radius-lg);
 }
