+++ apps/theme/components/Sidebar/Sidebar.module.css
@@ -95,7 +95,6 @@
   border-radius: 0 0 8px 8px;
   padding: 24px;
   border-top: 1px solid var(--ds-color-neutral-border-subtle);
-  background-color: var(--ds-color-info-background-default);
 }
 
 .appearance {
