+++ apps/theme/components/index.ts
@@ -11,8 +11,7 @@ export { ColorPreview } from './ColorPreview/ColorPreview';
 export { ColorPane } from './Sidebar/ColorPane/ColorPane';
 export { ColorTokens } from './ColorTokens/ColorTokens';
 export { ColorContrasts } from './ColorContrasts/ColorContrasts';
-export { BorderRadius } from './BorderRadius/BorderRadius';
-export { Theme1 } from './Previews/Theme1/Theme1';
+export { OverviewComponents } from './OverviewComponents/OverviewComponents';
 export { BorderRadiusInput } from './BorderRadiusInput/BorderRadiusInput';
 export { Header } from './Header/Header';
 export { ToggleChip } from './ToggleChip/ToggleChip';
